move out `typing.py`
